# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk Sistem Pengaduan \u0026 Tracking Fasilitas Kampus",
        "title": "SiLapor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/laporan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar laporan sesuai role user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laporan"
                ],
                "summary": "Daftar laporan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Laporan"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mahasiswa membuat laporan kerusakan baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laporan"
                ],
                "summary": "Buat laporan baru",
                "parameters": [
                    {
                        "description": "Data laporan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateLaporanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Laporan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/laporan/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail satu laporan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laporan"
                ],
                "summary": "Detail laporan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Laporan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Laporan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin menghapus laporan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laporan"
                ],
                "summary": "Hapus laporan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Laporan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/laporan/{id}/riwayat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Melihat timeline riwayat perubahan status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laporan"
                ],
                "summary": "Riwayat status laporan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Laporan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.RiwayatStatus"
                            }
                        }
                    }
                }
            }
        },
        "/api/laporan/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Petugas mengubah status laporan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laporan"
                ],
                "summary": "Ubah status laporan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Laporan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status baru",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Laporan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateLaporanRequest": {
            "type": "object",
            "required": [
                "deskripsi",
                "kategori_id",
                "lokasi"
            ],
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "foto_url": {
                    "type": "string"
                },
                "kategori_id": {
                    "type": "integer"
                },
                "lokasi": {
                    "type": "string"
                }
            }
        },
        "model.KategoriFasilitas": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nama_kategori": {
                    "type": "string"
                },
                "petugas": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.User"
                        }
                    ]
                },
                "petugas_id": {
                    "type": "integer"
                },
                "sla_jam": {
                    "type": "integer"
                }
            }
        },
...
        "model.Laporan": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "foto_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kategori": {
                    "$ref": "#/definitions/model.KategoriFasilitas"
                },
                "kategori_id": {
                    "type": "integer"
                },
                "lokasi": {
                    "type": "string"
                },
                "pelapor": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.User"
                        }
                    ]
                },
                "pelapor_id": {
                    "type": "integer"
                },
                "prioritas": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tanggal_lapor": {
                    "type": "string"
                },
                "tanggal_selesai": {
                    "type": "string"
                }
            }
        },
...
        "model.RiwayatStatus": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "keterangan": {
                    "type": "string"
                },
                "laporan": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Laporan"
                        }
                    ]
                },
                "laporan_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "waktu": {
                    "type": "string"
                }
            }
        },
        "model.UpdateStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}